Create README.md
In this project, we are using the RFID technology for identifying the books and
persons based on unique Tag number and Arduino Duemilanove board consisting of
Atmega328 Microcontroller which will process the data and a reader of 125 KHz
frequency (i.e., low frequency). Each student of the institution and each book in the
library will be attached a RFID tag. When a student wants a book to be issued, first
his/her Identity card is to be scanned and identified by the reader. Then, the books to
be issued are to be scanned and identified by the reader. With the help of
microcontroller and the program dumped into it, the books identified by the RFID
reader will be updated into the student’s account. At the end of the process, the count
of no. of books in a student’s account will be separately stored in the EEPROM
memory.